requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
supabase>=2.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
